changes the file names
