+++ src/adv.py
@@ -1,29 +1,33 @@
 from room import Room
 from player import Player
+from item import Item, Food, Egg
 # Declare all the rooms
-
 room = {
     'outside':  Room("Outside Cave Entrance",
                      "North of you, the cave mount beckons"),
-
     'foyer':    Room("Foyer", """Dim light filters in from the south. Dusty
 passages run north and east."""),
-
     'overlook': Room("Grand Overlook", """A steep cliff appears before you, falling
 into the darkness. Ahead to the north, a light flickers in
 the distance, but there is no way across the chasm."""),
-
     'narrow':   Room("Narrow Passage", """The narrow passage bends here from west
 to north. The smell of gold permeates the air."""),
-
     'treasure': Room("Treasure Chamber", """You've found the long-lost treasure
 chamber! Sadly, it has already been completely emptied by
 earlier adventurers. The only exit is to the south."""),
 }
 
+items = {
+    "sword": Item("sword", "A really old sword."),
+    "camera": Item("camera", "An old Leica...could be worth something!"),
+    "garbage": Item("garbage", "I'm in the middle of nowhere, how did garbage end up here?!")
+}
 
+# add items to rooms
+room["foyer"].items = [items["sword"]]
+room["overlook"].items = [items["camera"]]
+room["treasure"].items = [items["garbage"]]
 # Link rooms together
-
 room['outside'].n_to = room['foyer']
 room['foyer'].s_to = room['outside']
 room['foyer'].n_to = room['overlook']
@@ -33,70 +37,59 @@ room['narrow'].w_to = room['foyer']
 room['narrow'].n_to = room['treasure']
 room['treasure'].s_to = room['narrow']
 
+# creates items
+# rock = Item("rock", "This is a rock.")
+# sandwich = Food("sandwich", "This is a delicious sandwich.", 100)
+# egg = Egg()
 #
 # Main
 #
-
 # Make a new player object that is currently in the 'outside' room.
-# adventurer = Player("Colin", room["outside"])
-# active_room = adventurer.current_room
+player = Player(input("Please enter your name: "), room['outside'])
+print("You can go:", player.current_room)
+print("You are currently: ", player.current_room.name)
+print(player.current_room.description)
+# room["foyer"].items.append(rock)
+# room["outside"].items.append(rock)
+# player.items.append(rock)
+# player.items.append(sandwich)
+# print(player.take_item(rock))
+# player.take_item(rock)
+# player.items.append(egg)
+# player.eat(rock)
+# player.eat(sandwich)
+# player.eat(egg)
+# Write a loop that:
+#
+# * Prints the current room name
+# * Prints the current description (the textwrap module might be useful here).
+# * Waits for user input and decides what to do.
+#
+# If the user enters a cardinal direction, attempt to move to the room there.
+# Print an error message if the movement isn't allowed.
+#
+# If the user enters "q", quit the game.
 
-
-# for error handling so it is visible to the user
-def prRed(skk): print("\033[91m {}\033[00m" .format(skk))
-
-# for prompts at the beginning
-
-
-def prGreen(skk): print("\033[92m {}\033[00m" .format(skk))
-
-# for updating location prompts
-
-
-def prLightPurple(skk): print("\033[94m {}\033[00m" .format(skk))
-
-
-user_name = input("What is your name? ")
-player = Player(user_name, room["outside"])
-prGreen(f"Welcome, {player.name}!")
-prGreen("You can move by typing n, s, e, w to go in that cardinal direction.")
+valid_directions = ("n", "s", "e", "w")
 
 while True:
-    prLightPurple(
-        f">>> You are in: {player.current_room.name}\n>>> {player.current_room.description}")
-
-    cmd = input(f">>> {player.name} thinks... where to next? -> ")
-
-    if cmd == "n":
-        if player.current_room.n_to:
-            player.current_room = player.current_room.n_to
+    player.display_items()
+    cmd = input("\n~~> ")
+    if cmd == "q":
+        print("Goodbye!")
+        exit(0)
+    elif cmd in valid_directions:
+        player.travel(cmd)
+    elif cmd == "i":
+        player.print_inventory()
+    elif cmd.startswith("get") or cmd.startswith("take"):
+        cmd_word = cmd.split()
+        if len(player.current_room.items) > 0 and player.current_room.show_items(cmd_word[1]):
+            player.take_item(items[cmd_word[1]])
         else:
-            prRed("*************** There's no room to the North! ***************")
-    elif cmd == "s":
-        if player.current_room.s_to:
-            player.current_room = player.current_room.s_to
-        else:
-            prRed("*************** There's no room to the South! ***************")
-    elif cmd == "e":
-        if player.current_room.e_to:
-            player.current_room = player.current_room.e_to
-        else:
-            prRed("*************** There's no room to the East! ***************")
-    elif cmd == "w":
-        if player.current_room.w_to:
-            player.current_room = player.current_room.w_to
-        else:
-            prRed("*************** There's no room to the West! ***************")
-    elif cmd == "q":
-        quit()
+            print(f"This room doesn't have a {cmd_word[1]}")
+    elif cmd.startswith("drop"):
+        cmd_word = cmd.split()
+        player.drop_item(cmd_word[1])
     else:
-        prRed("That is no a valid entry, try n, s, w, e, or q to quit.")
-
-#         # * Prints the current room name
-#         # * Prints the current description (the textwrap module might be useful here).
-#         # * Waits for user input and decides what to do.
-#         #
-#         # If the user enters a cardinal direction, attempt to move to the room there.
-#         # Print an error message if the movement isn't allowed.
-#         #
-#         # If the user enters "q", quit the game.
+        print("I did not understand that command")
