+++ src/item_day1.py
@@ -1,0 +1,8 @@
+# Item Class
+# Items for now should be one word for easier processing
+
+
+class Item:
+    def __init__(self, name, description):
+        self.name = name
+        self.description = description
