+++ src/player_day1.py
@@ -1,0 +1,10 @@
+# Write a class to hold player information, e.g. what room they are in
+# currently.
+
+# Player Class
+
+
+class Player:
+    def __init__(self, name, current_room):
+        self.name = name
+        self.current_room = current_room
